requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Faker==20.1.0
numpy==1.24.3
tqdm==4.66.1